requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
pytest==8.3.4
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
